pyproject.toml
[project]
name = "openrouter-mcp"
version = "0.1.0"
description = "MCP server for OpenRouter API"
requires-python = ">=3.10"
dependencies = [
    "mcp>=0.1.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0"
]

[project.scripts]
openrouter-mcp = "openrouter_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/openrouter_mcp"]
